Normalized REST commonalities to a new library, normalized their use
