+++ lib/keys.js
@@ -3,7 +3,7 @@ var path = require('path')
     , rest = require('restler')
     , url = require('url')
     , utils = require('./utils')
-    , configtools = require('../lib/config')
+    , configtools = require('./config')
     ;
 module.exports = {
     getDefaultKey: function() {
@@ -25,50 +25,44 @@ module.exports = {
         return fs.readFileSync(path, 'utf8');
     },
     sendKey: function(key, callback) {
-        var baseUrl = configtools.getBaseUrl()
-            , sendUrl = url.resolve(baseUrl, 'auth/add-key')
+        var sendUrl = remotetools.getUrl('auth/add-key')
             ;
         if(process.env.DEXTER_DEV) {
             console.log('Sending key to', sendUrl);
         }
-        rest.post(sendUrl, {
+        rest.post(sendUrl, remotetools.signRequest({
             data: { key: key }
-            , headers: configtools.getAuthorizedRequestHeaders()
-        }).on('complete', function(result, response) {
+        })).on('complete', function(result, response) {
             utils.restResponse(result, response, function() {
                 if(callback) callback(result);
             });
         });
     },
     removeKey: function(query, callback) {
-        var baseUrl = configtools.getBaseUrl()
-            , sendUrl = url.resolve(baseUrl, 'auth/remove-key')
+        var removeUrl = remotetools.getUrl('auth/remove-key')
             ;
         if(process.env.DEXTER_DEV) {
-            console.log('Removing key:', sendUrl);
+            console.log('Removing key:', removeUrl);
         }
-        rest.del(sendUrl, {
+        rest.del(removeUrl, remotetools.signRequest({
             data: { query: query }
-            , headers: configtools.getAuthorizedRequestHeaders()
-        }).on('complete', function(result, response) {
+        })).on('complete', function(result, response) {
             utils.restResponse(result, response, function() {
                 if(callback) callback(result);
             });
         });
     },
     getAll: function(callback) {
-        var baseUrl = configtools.getBaseUrl()
-            , sendUrl = url.resolve(baseUrl, 'auth/list-keys')
+        var  getUrl = remotetools.getUrl('auth/list-keys')
             ;
         if(process.env.DEXTER_DEV) {
-            console.log('Fetching keys:', sendUrl);
+            console.log('Fetching keys:', getUrl);
         }
-        rest.get(sendUrl, {
-            headers: configtools.getAuthorizedRequestHeaders()
-        }).on('complete', function(result, response) {
-            utils.restResponse(result, response, function() {
-                if(callback) callback(result);
+        rest.get(getUrl, remotetools.signRequest())
+            .on('complete', function(result, response) {
+                utils.restResponse(result, response, function() {
+                    if(callback) callback(result);
+                });
             });
-        });
     }
 }
