+++ lib/utils.js
@@ -40,27 +40,5 @@ module.exports = {
         process.stdin.on('end', function() {
             cb(content);
         });
-    },
-    restResponse: function(result, response, success) {
-        if(result && result.success) {
-            success();
-        } else if(process.env.DEXTER_DEV) {
-            if(response && response.statusCode && response.statusCode) {
-                console.error(response.statusCode);
-                if(response.statusCode != 404) {
-                    console.log(result);
-                }
-            } else {
-                console.error(response);
-            }
-        } else {
-            if(result && result.error) {
-                console.error(result.error);
-            } else if(response && response.statusCode) {
-                console.error(response.statusCode, result);
-            } else {
-                console.log(result.code);
-            }
-        }
     }
 }
